Extract command-collection-related functions
I'm just experimenting with this.. But I find it confusing to keep
command and command-collection in the same module. The naming gets
weird.
