+++ lib/chukinas/dreadnought/command.ex
@@ -26,44 +26,22 @@ defmodule Chukinas.Dreadnought.Command do
   end
 
   # *** *******************************
-  # *** COMMAND OPERATIONS
-
-  # TODO make note about how segments are listed from later (future) to oldest (past)
-  defp get_segment_numbers(command) do
-    min_segment = command.segment_number
-    max_segment = command.segment_number + command.segment_count - 1
-    min_segment..max_segment
-  end
-
-  defp occupies_segment(command, segment_number) do
-    segment_number in get_segment_numbers(command)
-  end
-
-  # *** *******************************
-  # *** LIST OPERATIONS
-
-  def remove_at_segment(commands, segment_number) when is_list(commands) do
-    {before_commands, [removed_command, after_commands]} = commands
-    |> Enum.split_while(&(!occupies_segment(&1, segment_number)))
-    new_commands = get_segment_numbers(removed_command) |> Enum.map(&new/1)
-    Enum.concat([before_commands, new_commands, after_commands])
-  end
-
-  def set_path(commands, start_vector) when is_list(commands) do
-    # flat map reduce
-    IO.inspect(start_vector)
-    {commands_with_paths, _final_end_vector} = commands
-    |> Enum.map_reduce(start_vector, fn x, acc ->
-      command = set_path(x, acc) |> IO.inspect()
-      {command, command.vector_end}
-    end)
-    commands_with_paths
-  end
+  # *** API
 
   def set_path(%__MODULE__{angle: 0, speed: 3} = command, %Vector{} = start_vector) do
     command
     |> Map.put(:vector_start, start_vector)
     |> Map.put(:vector_end, Vector.move_straight(start_vector, 100))
+  end
+
+  # TODO make note about how segments are listed from later (future) to oldest (past)
+  def get_segment_numbers(%__MODULE__{segment_number: min_segment} = command) do
+    max_segment = min_segment + command.segment_count - 1
+    min_segment..max_segment
+  end
+
+  def occupies_segment(%__MODULE__{} = command, segment_number) do
+    segment_number in get_segment_numbers(command)
   end
 
 end
